this still fails, I guess I need to actually think about this...

+++ TinyWireS/usiTwiSlave.c
@@ -404,7 +404,7 @@ usiTwiDataInReceiveBuffer(
 
 uint8_t usiTwiAmountDataInReceiveBuffer(void)
 {
-    return abs(rxTail - rxHead);
+    return abs((int)rxTail - (int)rxHead);
 }
